Safari-611.1.10/LayoutTests/http/tests/security/xss-DENIED-method-with-iframe-proto-expected.txt

CONSOLE MESSAGE: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.

Tests that making other frame window a prototype doesn't expose that window methods

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS __proto__ = targetWindow threw exception TypeError: Object.prototype.__proto__ called on null or undefined.
PASS: successfullyParsed should be 'true' and is.

TEST COMPLETE
